+++ src/index.ts
@@ -7,3 +7,4 @@ export * from "./components/utilities/MapOrFallback.js";
 export * from "./components/Block.js";
 export * from "./components/Breadcrumbs.js";
 export * from "./components/Header.js";
+export * from "./components/Notice.js";